pyproject.toml
[project]
name = "bn"
version = "0.1.0"
description = "Multi Language Invoice Extractor LLm Project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.0.9",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
